Combined Josh's DMX changes for running the show with the issue14 changes
