+++ app/html/dmx.html
@@ -59,6 +59,6 @@
     <br>
 
 </body>
-    <script type="text/javascript" src="../js/dmx_output.js"></script>
+    <script type="text/javascript" src="../js/dmxOutput.js"></script>
     <script type="text/javascript" src="../js/dmx.js"></script>
 </html>
