+++ app/html/show.html
@@ -35,7 +35,5 @@
         var onlyShowName = iframe.value.split(path.sep);
         onlyShowName = iframe.value.split(path.sep)[onlyShowName.length - 1];
         header.innerText = onlyShowName;
-        
-
     </script> 
 </html>
